--- Example/KeyboardSupport/AppDelegate.swift
@@ -1,18 +1,0 @@
-//
-//  AppDelegate.swift
-//  KeyboardSupport_Example
-//
-//  Copyright (c) 2017 Bottle Rocket Studios. All rights reserved.
-//
-
-import UIKit
-
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    var window: UIWindow?
-
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
-        return true
-    }
-}
