--- Example/KeyboardSupport/CollectionViewCell.swift
@@ -1,13 +1,0 @@
-//
-//  CollectionViewCell.swift
-//  KeyboardSupport_Example
-//
-//  Copyright © 2017 CocoaPods. All rights reserved.
-//
-
-import UIKit
-
-/// A simple cell with a text field.
-class CollectionViewCell: UICollectionViewCell {
-    @IBOutlet private var textField: UITextField!
-}
