+++ Examples/KeyboardSupport-iOSExample/AppDelegate.swift
@@ -1,0 +1,19 @@
+//
+//  AppDelegate.swift
+//  KeyboardSupport-iOSExample
+//
+//  Created by Earl Gaspard on 5/28/18.
+//  Copyright © 2018 Bottle Rocket Studios. All rights reserved.
+//
+
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    var window: UIWindow?
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
+        return true
+    }
+}
