+++ Examples/KeyboardSupport-iOSExample/CollectionViewCell.swift
@@ -1,0 +1,14 @@
+//
+//  CollectionViewCell.swift
+//  KeyboardSupport_Example
+//
+//  Created by Earl Gaspard on 12/19/17.
+//  Copyright © 2017 Bottle Rocket Studios. All rights reserved.
+//
+
+import UIKit
+
+/// A simple cell with a text field.
+class CollectionViewCell: UICollectionViewCell {
+    @IBOutlet private var textField: UITextField!
+}
