+++ Sources/Supporting Files/KeyboardSupport.h
@@ -1,0 +1,19 @@
+//
+//  KeyboardSupport_iOS.h
+//  KeyboardSupport-iOS
+//
+//  Created by Earl Gaspard on 5/28/18.
+//  Copyright © 2018 Bottle Rocket. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for KeyboardSupport_iOS.
+FOUNDATION_EXPORT double KeyboardSupport_iOSVersionNumber;
+
+//! Project version string for KeyboardSupport_iOS.
+FOUNDATION_EXPORT const unsigned char KeyboardSupport_iOSVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <KeyboardSupport_iOS/PublicHeader.h>
+
+
